pyproject.toml
[tool.poetry]
name = "detemineendableautomate"
version = "0.1.0"
description = ""
authors = ["Цвиркунов Михаил <tsvirkunov@profilum.ru>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
poetry = "^1.8.3"
pytest = "^8.3.3"
pydantic = "^2.9.2"
pytest-mock = "^3.14.0"
pytest-bdd = "^7.3.0"
mock = "^5.1.0"
jax = "^0.4.33"
behave = "^1.2.6"
polars = "^1.9.0"
coverage = "^7.6.3"
python-ioc = "^1.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "src"
]
pythonpath = "src"
bdd_features_base_dir = "tests/features"

# [tool.coverage.report]
# exclude_also = [

# ]
